initial tests implemented
