+++ 05_CleanCode/GBShop/GBShop/App/BusinessLogic/Network/ChangeUserDataRequestFactory.swift
@@ -16,5 +16,5 @@ protocol ChangeUserDataRequestFactory {
                 gender: String,
                 creditCard: String,
                 bio: String,
-                completionHandler: @escaping (AFDataResponse<ResponseResult>) -> Void)
+                completionHandler: @escaping (AFDataResponse<CommonResponseResult>) -> Void)
 }
