+++ 05_CleanCode/GBShop/GBShop/Entity/Catalog.swift
@@ -1,0 +1,18 @@
+//
+//  Catalog.swift
+//  GBShop
+//
+//  Created by Evgeny Alekseev on 11.04.2022.
+//
+
+import Foundation
+
+struct Catalog: Codable {
+    let pageNumber: Int
+    let products: [Product?]
+    
+    enum CodingKeys: String, CodingKey {
+        case pageNumber = "page_number"
+        case products
+    }
+}
