+++ 05_CleanCode/GBShop/GBShop/Entity/Product.swift
@@ -1,5 +1,5 @@
 //
-//  SalesItem.swift
+//  Product.swift
 //  GBShop
 //
 //  Created by Evgeny Alekseev on 11.04.2022.
@@ -9,12 +9,14 @@ import Foundation
 
 struct Product: Codable {
     let result: Int
+    let productID: Int?
     let productName: String
     let productPrice: Double
-    let productDescription: String
+    let productDescription: String?
     
     enum CodingKeys: String, CodingKey {
         case result
+        case productID = "id_product"
         case productName = "product_name"
         case productPrice = "product_price"
         case productDescription = "product_description"
