--- 05_CleanCode/GBShop/GBShop/Entity/ResponseResult.swift
@@ -1,21 +1,0 @@
-//
-//  LoginResult.swift
-//  GBShop
-//
-//  Created by Evgeny Alekseev on 20.03.2022.
-//
-
-import Foundation
-
-struct ResponseResult: Codable {
-    let result: Int
-    let user: User?
-    let authToken: String?
-    let userMessage: String?
-    let errorMessage: String?
-}
-
-struct ProductResponseResult: Codable {
-    let result: Int
-    let 
-}
