+++ 05_CleanCode/GBShop/GBShopTests/ResponseTests.swift
@@ -2,78 +2,106 @@
 //  ResponseTests.swift
 //  GBShopTests
 //
-//  Created by Evgeny Alekseev on 27.03.2022.
+//  Created by Evgeny Alekseev on 14.03.2022.
 //
 
 import XCTest
-import Alamofire
 @testable import GBShop
-
-struct PostStub: Codable {
-    let result: Int
-}
-
-enum ApiErrorStub: Error {
-    case fatalError
-}
-
-struct ErrorParserStub: AbstractErrorParser {
-    func parse(_ result: Error) -> Error {
-        return ApiErrorStub.fatalError
-    }
-    
-    func parse(response: HTTPURLResponse?,
-               data: Data?,
-               error: Error?) -> Error? {
-        return error
-    }
-}
 
 class ResponseTests: XCTestCase {
     
-    let expectation = XCTestExpectation(description: "Download https://failUrl")
-    var errorParser: ErrorParserStub!
+    var requestFactory: RequestFactory!
+    var waitingTime: TimeInterval!
     
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
-        errorParser = ErrorParserStub()
+        requestFactory = RequestFactory()
+        waitingTime = 10
     }
     
     override func tearDownWithError() throws {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
-        errorParser = nil
+        requestFactory = nil
+        waitingTime = nil
     }
     
-    func test_shouldAuthorize() {
+    // MARK: test auth request
+    func test_loginRequest() {
+        let request = requestFactory.makeAuthRequestFactory()
+        let expectation = expectation(description: "Login complete")
         
-        AF.request("https://raw.githubusercontent.com/GeekBrainsTutorial/online-store-api/master/responses/login.json")
-            .responseCodable(errorParser: errorParser) { [weak self] (response: DataResponse<PostStub, AFError>) in
-                switch response.result {
-                case .success(let result):
-                    print("*** Result: \(result) ***")
-                    break
-                case .failure(let error):
-                    print("*** Error: \(error) ***")
-                    XCTFail()
-                }
-                self?.expectation.fulfill()
+        request.login(username: "Somebody", password: "mypassword" ) { response in
+            switch response.result {
+            case .success(let result):
+                print("*** Login result: \(result) ***")
+            case .failure:
+                XCTFail()
             }
-        wait(for: [expectation], timeout: 10.0)
-    }
-    
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        // Any test you write for XCTest can be annotated as throws and async.
-        // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
-        // Mark your test async to allow awaiting for asynchronous code to complete. Check the results with assertions afterwards.
-    }
-    
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
+            expectation.fulfill()
         }
+        waitForExpectations(timeout: waitingTime)
     }
     
+    // MARK: test logout request
+    func test_logoutRequest() {
+        let request = requestFactory.makeLogoutRequestFactory()
+        let expectation = expectation(description: "Logout complete")
+        
+        request.logout(userID: 1) { response in
+            switch response.result {
+            case .success(let result):
+                print("*** Logout result: \(result) ***")
+            case .failure:
+                XCTFail()
+            }
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: waitingTime)
+    }
+    
+    // MARK: test registration request
+    func test_registrationRequest() {
+        let request = requestFactory.makeRegisterUserRequestFactory()
+        let expectation = expectation(description: "Registration complete")
+        
+        request.register(userID: 2,
+                         username: "User2",
+                         password: "password2",
+                         email: "user2@test.com",
+                         gender: "m",
+                         creditCard: "1111222233334444",
+                         bio: "I am Groot") { response in
+            switch response.result {
+            case .success(let result):
+                print("*** Registration result: \(result) ***")
+            case .failure:
+                XCTFail()
+            }
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: waitingTime)
+    }
+    
+    // MARK: test change user data
+    func test_changeUserDataRequest() {
+        let request = requestFactory.makeChangeUserDataRequestFactory()
+        let expectation = expectation(description: "Data changed")
+        
+        request.change(userID: 2,
+                       username: "newUserName",
+                       password: "newPassword",
+                       email: "newEmail@",
+                       gender: "f",
+                       creditCard: "9999888877776666",
+                       bio: "I am not Groot") { response in
+            switch response.result {
+            case .success(let result):
+                print("*** Change user data result: \(result) ***")
+            case .failure:
+                XCTFail()
+            }
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: waitingTime)
+    }
 }
